# Optimize MIR expression data structures