--- a/lib/mir/src/ir/byte_string.rs
+++ b/lib/mir/src/ir/byte_string.rs
@@ -1,12 +1,14 @@
+use std::sync::Arc;
+
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct ByteString {
-    value: Vec<u8>,
+    value: Arc<[u8]>,
 }
 
 impl ByteString {
     pub fn new(value: impl Into<Vec<u8>>) -> Self {
         Self {
-            value: value.into(),
+            value: value.into().into(),
         }
     }