--- a/lib/mir/src/ir/call.rs
+++ b/lib/mir/src/ir/call.rs
@@ -3,9 +3,12 @@ use crate::types;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Call {
+pub struct Call(Arc<CallInner>);
+
+#[derive(Debug, PartialEq)]
+struct CallInner {
     type_: types::Function,
-    function: Arc<Expression>,
+    function: Expression,
     arguments: Vec<Expression>,
 }
 
@@ -15,22 +18,25 @@ impl Call {
         function: impl Into<Expression>,
         arguments: Vec<Expression>,
     ) -> Self {
-        Self {
-            type_,
-            function: function.into().into(),
-            arguments,
-        }
+        Self(
+            CallInner {
+                type_,
+                function: function.into().into(),
+                arguments,
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> &types::Function {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn function(&self) -> &Expression {
-        &self.function
+        &self.0.function
     }
 
     pub fn arguments(&self) -> &[Expression] {
-        &self.arguments
+        &self.0.arguments
     }
 }