--- a/lib/mir/src/ir/comparison_operation.rs
+++ b/lib/mir/src/ir/comparison_operation.rs
@@ -2,10 +2,13 @@ use super::{comparison_operator::ComparisonOperator, expression::Expression};
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct ComparisonOperation {
+pub struct ComparisonOperation(Arc<ComparisonOperationInner>);
+
+#[derive(Debug, PartialEq)]
+struct ComparisonOperationInner {
     operator: ComparisonOperator,
-    lhs: Arc<Expression>,
-    rhs: Arc<Expression>,
+    lhs: Expression,
+    rhs: Expression,
 }
 
 impl ComparisonOperation {
@@ -14,22 +17,25 @@ impl ComparisonOperation {
         lhs: impl Into<Expression>,
         rhs: impl Into<Expression>,
     ) -> Self {
-        Self {
-            operator,
-            lhs: Arc::new(lhs.into()),
-            rhs: Arc::new(rhs.into()),
-        }
+        Self(
+            ComparisonOperationInner {
+                operator,
+                lhs: (lhs.into()),
+                rhs: (rhs.into()),
+            }
+            .into(),
+        )
     }
 
     pub fn operator(&self) -> ComparisonOperator {
-        self.operator
+        self.0.operator
     }
 
     pub fn lhs(&self) -> &Expression {
-        &self.lhs
+        &self.0.lhs
     }
 
     pub fn rhs(&self) -> &Expression {
-        &self.rhs
+        &self.0.rhs
     }
 }