--- a/lib/mir/src/ir/comparison_operation.rs
+++ b/lib/mir/src/ir/comparison_operation.rs
@@ -2,7 +2,7 @@ use super::{comparison_operator::ComparisonOperator, expression::Expression};
 use std::rc::Rc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct ComparisonOperation(Arc<ComparisonOperationInner>);
+pub struct ComparisonOperation(Rc<ComparisonOperationInner>);
 
 #[derive(Debug, PartialEq)]
 struct ComparisonOperationInner {