--- a/lib/mir/src/ir/drop_variables.rs
+++ b/lib/mir/src/ir/drop_variables.rs
@@ -4,24 +4,30 @@ use fnv::FnvHashMap;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct DropVariables {
+pub struct DropVariables(Arc<DropVariablesInner>);
+
+#[derive(Debug, PartialEq)]
+struct DropVariablesInner {
     variables: FnvHashMap<String, Type>,
-    expression: Arc<Expression>,
+    expression: Expression,
 }
 
 impl DropVariables {
     pub fn new(variables: FnvHashMap<String, Type>, expression: impl Into<Expression>) -> Self {
-        Self {
-            variables,
-            expression: expression.into().into(),
-        }
+        Self(
+            DropVariablesInner {
+                variables,
+                expression: expression.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn variables(&self) -> &FnvHashMap<String, Type> {
-        &self.variables
+        &self.0.variables
     }
 
     pub fn expression(&self) -> &Expression {
-        &self.expression
+        &self.0.expression
     }
 }