--- a/lib/mir/src/ir/drop_variables.rs
+++ b/lib/mir/src/ir/drop_variables.rs
@@ -17,7 +17,7 @@ impl DropVariables {
         Self(
             DropVariablesInner {
                 variables,
-                expression: expression.into().into(),
+                expression: expression.into(),
             }
             .into(),
         )