--- a/lib/mir/src/ir/if_.rs
+++ b/lib/mir/src/ir/if_.rs
@@ -2,10 +2,13 @@ use super::expression::Expression;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct If {
-    condition: Arc<Expression>,
-    then: Arc<Expression>,
-    else_: Arc<Expression>,
+pub struct If(Arc<IfInner>);
+
+#[derive(Clone, Debug, PartialEq)]
+struct IfInner {
+    condition: Expression,
+    then: Expression,
+    else_: Expression,
 }
 
 impl If {
@@ -14,22 +17,25 @@ impl If {
         then: impl Into<Expression>,
         else_: impl Into<Expression>,
     ) -> Self {
-        Self {
-            condition: Arc::new(condition.into()),
-            then: Arc::new(then.into()),
-            else_: Arc::new(else_.into()),
-        }
+        Self(
+            IfInner {
+                condition: (condition.into()),
+                then: (then.into()),
+                else_: (else_.into()),
+            }
+            .into(),
+        )
     }
 
     pub fn condition(&self) -> &Expression {
-        &self.condition
+        &self.0.condition
     }
 
     pub fn then(&self) -> &Expression {
-        &self.then
+        &self.0.then
     }
 
     pub fn else_(&self) -> &Expression {
-        &self.else_
+        &self.0.else_
     }
 }