--- a/lib/mir/src/ir/let_.rs
+++ b/lib/mir/src/ir/let_.rs
@@ -3,11 +3,14 @@ use crate::types::Type;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Let {
+pub struct Let(Arc<LetInner>);
+
+#[derive(Clone, Debug, PartialEq)]
+struct LetInner {
     name: String,
     type_: Type,
-    bound_expression: Arc<Expression>,
-    expression: Arc<Expression>,
+    bound_expression: Expression,
+    expression: Expression,
 }
 
 impl Let {
@@ -17,27 +20,30 @@ impl Let {
         bound_expression: impl Into<Expression>,
         expression: impl Into<Expression>,
     ) -> Self {
-        Self {
-            name: name.into(),
-            type_: type_.into(),
-            bound_expression: bound_expression.into().into(),
-            expression: expression.into().into(),
-        }
+        Self(
+            LetInner {
+                name: name.into(),
+                type_: type_.into(),
+                bound_expression: bound_expression.into().into(),
+                expression: expression.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn name(&self) -> &str {
-        &self.name
+        &self.0.name
     }
 
     pub fn type_(&self) -> &Type {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn bound_expression(&self) -> &Expression {
-        &self.bound_expression
+        &self.0.bound_expression
     }
 
     pub fn expression(&self) -> &Expression {
-        &self.expression
+        &self.0.expression
     }
 }