--- a/lib/mir/src/ir/record.rs
+++ b/lib/mir/src/ir/record.rs
@@ -1,22 +1,26 @@
 use super::expression::Expression;
 use crate::types;
+use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Record {
+pub struct Record(Arc<RecordInner>);
+
+#[derive(Debug, PartialEq)]
+struct RecordInner {
     type_: types::Record,
     fields: Vec<Expression>,
 }
 
 impl Record {
     pub fn new(type_: types::Record, fields: Vec<Expression>) -> Self {
-        Self { type_, fields }
+        Self(RecordInner { type_, fields }.into())
     }
 
     pub fn type_(&self) -> &types::Record {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn fields(&self) -> &[Expression] {
-        &self.fields
+        &self.0.fields
     }
 }