--- a/lib/mir/src/ir/record_field.rs
+++ b/lib/mir/src/ir/record_field.rs
@@ -3,30 +3,36 @@ use crate::types;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct RecordField {
+pub struct RecordField(Arc<RecordFieldInner>);
+
+#[derive(Debug, PartialEq)]
+struct RecordFieldInner {
     type_: types::Record,
     index: usize,
-    record: Arc<Expression>,
+    record: Expression,
 }
 
 impl RecordField {
     pub fn new(type_: types::Record, index: usize, record: impl Into<Expression>) -> Self {
-        Self {
-            type_,
-            index,
-            record: record.into().into(),
-        }
+        Self(
+            RecordFieldInner {
+                type_,
+                index,
+                record: record.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> &types::Record {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn index(&self) -> usize {
-        self.index
+        self.0.index
     }
 
     pub fn record(&self) -> &Expression {
-        &self.record
+        &self.0.record
     }
 }