--- a/lib/mir/src/ir/record_field.rs
+++ b/lib/mir/src/ir/record_field.rs
@@ -18,7 +18,7 @@ impl RecordField {
             RecordFieldInner {
                 type_,
                 index,
-                record: record.into().into(),
+                record: record.into(),
             }
             .into(),
         )