--- a/lib/mir/src/ir/record_field.rs
+++ b/lib/mir/src/ir/record_field.rs
@@ -3,7 +3,7 @@ use crate::types;
 use std::rc::Rc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct RecordField(Arc<RecordFieldInner>);
+pub struct RecordField(Rc<RecordFieldInner>);
 
 #[derive(Debug, PartialEq)]
 struct RecordFieldInner {