--- a/lib/mir/src/ir/string_concatenation.rs
+++ b/lib/mir/src/ir/string_concatenation.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct StringConcatenation {
-    operands: Arc<[Expression]>,
+    operands: Rc<[Expression]>,
 }
 
 impl StringConcatenation {