--- a/lib/mir/src/ir/synchronize.rs
+++ b/lib/mir/src/ir/synchronize.rs
@@ -3,24 +3,30 @@ use crate::types::Type;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Synchronize {
+pub struct Synchronize(Arc<SynchronizeInner>);
+
+#[derive(Debug, PartialEq)]
+struct SynchronizeInner {
     type_: Type,
-    expression: Arc<Expression>,
+    expression: Expression,
 }
 
 impl Synchronize {
     pub fn new(type_: impl Into<Type>, expression: impl Into<Expression>) -> Self {
-        Self {
-            type_: type_.into(),
-            expression: expression.into().into(),
-        }
+        Self(
+            SynchronizeInner {
+                type_: type_.into(),
+                expression: expression.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> &Type {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn expression(&self) -> &Expression {
-        &self.expression
+        &self.0.expression
     }
 }