--- a/lib/mir/src/ir/try_operation.rs
+++ b/lib/mir/src/ir/try_operation.rs
@@ -5,11 +5,14 @@ use std::sync::Arc;
 // A try operation matches an operand with a type and returns it from a function
 // if it matches.
 #[derive(Clone, Debug, PartialEq)]
-pub struct TryOperation {
-    operand: Arc<Expression>,
+pub struct TryOperation(Arc<TryOperationInner>);
+
+#[derive(Debug, PartialEq)]
+struct TryOperationInner {
+    operand: Expression,
     name: String,
     type_: Type,
-    then: Arc<Expression>,
+    then: Expression,
 }
 
 impl TryOperation {
@@ -19,27 +22,30 @@ impl TryOperation {
         type_: impl Into<Type>,
         then: impl Into<Expression>,
     ) -> Self {
-        Self {
-            operand: operand.into().into(),
-            name: name.into(),
-            type_: type_.into(),
-            then: then.into().into(),
-        }
+        Self(
+            TryOperationInner {
+                operand: operand.into().into(),
+                name: name.into(),
+                type_: type_.into(),
+                then: then.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn operand(&self) -> &Expression {
-        &self.operand
+        &self.0.operand
     }
 
     pub fn name(&self) -> &str {
-        &self.name
+        &self.0.name
     }
 
     pub fn type_(&self) -> &Type {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn then(&self) -> &Expression {
-        &self.then
+        &self.0.then
     }
 }