--- a/lib/mir/src/ir/variant.rs
+++ b/lib/mir/src/ir/variant.rs
@@ -3,24 +3,30 @@ use crate::types::Type;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Variant {
+pub struct Variant(Arc<VariantInner>);
+
+#[derive(Debug, PartialEq)]
+struct VariantInner {
     type_: Type,
-    payload: Arc<Expression>,
+    payload: Expression,
 }
 
 impl Variant {
     pub fn new(type_: impl Into<Type>, payload: impl Into<Expression>) -> Self {
-        Self {
-            type_: type_.into(),
-            payload: payload.into().into(),
-        }
+        Self(
+            VariantInner {
+                type_: type_.into(),
+                payload: payload.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> &Type {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn payload(&self) -> &Expression {
-        &self.payload
+        &self.0.payload
     }
 }