--- a/lib/mir/src/ir/variant.rs
+++ b/lib/mir/src/ir/variant.rs
@@ -16,7 +16,7 @@ impl Variant {
         Self(
             VariantInner {
                 type_: type_.into(),
-                payload: payload.into().into(),
+                payload: payload.into(),
             }
             .into(),
         )