--- a/lib/mir/src/types/type_.rs
+++ b/lib/mir/src/types/type_.rs
@@ -38,3 +38,14 @@ impl From<Record> for Type {
         Self::Record(record)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::mem::size_of;
+
+    #[test]
+    fn type_size() {
+        assert_eq!(size_of::<Type>(), 2 * size_of::<usize>());
+    }
+}